requirements.txt
Pillow>=10.0.0
aiohttp>=3.9.0